+++ rust/services/chain/server_lib/tests/integration_tests.rs
@@ -7,7 +7,7 @@ use server_utils::{body_to_json, post};
 
 #[tokio::test]
 async fn http_not_found() {
-    let db = ChainDb::in_memory(GuestElf::default());
+    let db = ChainDb::in_memory([GuestElf::default().id]);
     let app = server(db);
     let empty_body = json!({});
     let response = post(app, "/non-existent", &empty_body).await;
@@ -16,7 +16,7 @@ async fn http_not_found() {
 
 #[tokio::test]
 async fn method_not_found() {
-    let db = ChainDb::in_memory(GuestElf::default());
+    let db = ChainDb::in_memory([GuestElf::default().id]);
     let app = server(db);
     let req = json!({
         "jsonrpc": "2.0",
@@ -42,7 +42,7 @@ async fn method_not_found() {
 
 #[tokio::test]
 async fn method_missing() {
-    let db = ChainDb::in_memory(GuestElf::default());
+    let db = ChainDb::in_memory([GuestElf::default().id]);
     let app = server(db);
     let req = json!({
         "jsonrpc": "2.0",
@@ -71,7 +71,7 @@ mod chain_proof {
     #[tokio::test]
     #[ignore]
     async fn success_dummy() {
-        let db = ChainDb::in_memory(GuestElf::default());
+        let db = ChainDb::in_memory([GuestElf::default().id]);
         let app = server(db);
         let req = json!({
             "jsonrpc": "2.0",
@@ -100,7 +100,7 @@ mod chain_proof {
 
     #[tokio::test]
     async fn no_block_numbers_error() {
-        let db = ChainDb::in_memory(GuestElf::default());
+        let db = ChainDb::in_memory([GuestElf::default().id]);
         let app = server(db);
         let req = json!({
             "jsonrpc": "2.0",
@@ -129,7 +129,7 @@ mod chain_proof {
 
     #[tokio::test]
     async fn field_validation_error() {
-        let db = ChainDb::in_memory(GuestElf::default());
+        let db = ChainDb::in_memory([GuestElf::default().id]);
         let app = server(db);
 
         let valid_number = 42;