+++ rust/guest_wrapper/artifacts/chain_guest/CHANGELOG.md
@@ -3,3 +3,4 @@
   * `e8175d54bac30389a87a8dd39ac43fe22f621066d4e657eee10a30328bdfb05f` – Downgrade revm from 19.2.0 to 19.0.0 [#1557](https://github.com/vlayer-xyz/vlayer/pull/1557)
   * `ed5b4117fe05afa4f5136dbbac0e3a6c0a51394715e5f7fec64cf2dfd6319fa1` – Risc0 v1.2.1 (with accelerated keccaks) [#1551](https://github.com/vlayer-xyz/vlayer/pull/1551)
   * `2dfd3adf61327dc1ad47a637f5771496458ad5dc9a10de41b05cb348fa5d2eee` – Use old guest IDs in chain guest [#1584](https://github.com/vlayer-xyz/vlayer/pull/1584)
+  * `acbedbbeb664cdc1c58abd3d1c6c5a6e66d7cd206d752e3ddb7bf133f5f04ce7` – Use old guest IDs in chain host [#1603](https://github.com/vlayer-xyz/vlayer/pull/1603)